:fire: [10] REMOVE detailpage and resolve it in ZaaktypenView
